Handle multiple result run output parameters equally
Some values are always displayed in the tabular result even if they
never change when the program is run. For example, LAP. With this
change, all values that do not change during the run are inserted to
the static parameter values table instead.

All tabulated values are now treated equally, without special cases.