Allocate memory prior to starting calculations
Allocate all models and assign values prior to initiating the
calculation step. This allows for OOM errors to be caught before
significant amount of time is invested in calculations.

The model and integration helpers were modified to throw bad_alloc
exception in event that memory allocation fails. OpenCL scratch space
is now allocated at initialization of OpenCL devices, not in each
integration step reducing pressure on the memory allocater.