Warn if model will use more than 75% of RAM and 1m models
Since it is not a good idea to kill user's desktop by initiating
massive swapping when calculation is to be done. Warn user when more
than 75% of RAM is to be used by a model and outright refuse to run if
total memory consumption would exceed total memory space.

Also limit number of models that are to be calculated to below
1,000,000.