Double lung model and model calibration
Added double lung model. In the double lung model, the first
bifurcation is at the same GP level. Then each "half" of the vessels
are assigned the same GP as the other half - 2 parallel lungs. This
allows for disease simulation in only one of the 2 lungs.

Also added calibration window. This allows to calibrate the model to
specific "normal" physiological parameters. The default model is
calibrated to 42%/16%/42% resistance ratios (Rus/PVR,Rm/PVR,Rds/PVR)
with standard parameters as in the old model.

Calibration values are stored in local setting file.