Add boundary check for disease parameters
When entering a disease parameter value, bound the min/max values to
the range values as determined by the disease's parameters() function.

Range values that are outside the range values of the parameters()
function are discarded as these should not happen in the first
place. The RangeStepDlg already caps range definitions so only manual
range entry can trigger this behaviour.