Change the model from resistance-based to diameter-based
The former model used resistances as the baseline for calculation of
solution given flow and pressures. This led to some unrealistic vessel
sizes despite it being consistent.

The model was modified such that vessel diameter ratios and length
ratios are as per W. Huang et al. paper "Morphometry of the human
pulmonary vasculature". The model still uses simple biforcation model
and does not use connectivity matrix, but the results are now more
consistent with real life measurements (at least to vessel
diameters). This allows for blood viscosity to be included in the
model calculations.