Range::contains() failed when one of ranges positions were 0 or negative
The comparison with > and < while scaling range [min,max] caused
problems when one of the values was 0. In this case, we had,

  if (0 > -0)

which is clearly false while the intention is clearly that the
comparison is true. So now we detect 0 case and scale properly in case
of range being negative.