Add support for completely closed vessels (D=0)
Closed vessels present a little bit of a problem to the numerical
simulation as they introduce infinities in resistances and using
standard flow/pressure formulaes results in undefined numbers. Further
attempts at using these results in undefined.

So, it is important to detect and handle this important edge
case. Closed vessel means infinite resistance and no flow. Pressure
across the vessel is then undefined and should be ignored, as if the
vessel does not exist.