Added overlay map dialog with a flattened overlay
In the main model, the lung is horizontally and vertically compressed,
such that each successive generation occupies 1/2 the height and width
of a previous generation vessels. This is to allow proper scaling of
the vessel text areas, and generally to have correct scaling factor
for the graphical representation.

But with an overlay, it becomes difficult to see detail in the
inner-generations when zoomed out, something that an overlay should be
conveying. This is what the OverlayMapWidget addresses. It displays
each generation at the same width in a fixed, grid layout.