Add vessel ID string to each vessel drawn
Add title of the format "Artery Rt. (n of total)"
where n is index from bottom of the lung, and total is number in a
given lung. The dual lung model would have total*2 number of vessels
in a given generation.

Also removed units from viscosity factor, is it is dimensionless -
just a multiple of baseline 3.2cP blood viscosity.