Add overlay tooltip value display box and set range to 2 stddev
The OverlayMapWidget how displays a "tooltip" widget with value
information for the moused-over area. The value is limited to the
scale used in the image, hence the scale was increased to 2 standard
deviations instead of just 1. This allows more information to be
available and the readability of the gradient mostly remains.