Add vessel profile dialog and change artery (D/D0)(Ptp) function
The linear nature of the D/D0 relationship for arteries is not
possible - vessels to do expand to infinity. So adapt a function that
looks like the asymptotic vein function but expand the "linear" region
to Ptp of 70 as per the linear relationship. Linear relationship was
measured to be linear when Ptp is in 20-55 range. This will require
further review.

Vessel shapes are now drawn in the vessel detail window along with all
the calculated parameters using raw units.