Move GP handling from the integration routine to pressure/flow calculation
Replaced
  vessel.pressure
with
  vessel.pressure_in
  vessel.pressure_out

The old method of specifying pressure was a little confusing. It
apparently was equivalent to pressure_out for arteries, and
pressure_in for veins. That is, the old pressure was always "towards
the inside of the model". This caused confusion in the way pressure
was calculated and displayed.

Since each vessel now has pressure values at both its beginning and
end, it is simple to account for gravitational potential (GP) as GP is
just static part of the model geometry. Each bifurcation has a change
in GP, hence a corresponding change in pressure.

  pressure_in[new] = pressure_out[old] - (GP_new - GP_old)

This greatly simplifies handling of GP as now it only appears in
pressure calculation function, Model::calculateChildrenFlowPress()