Avoid condition where floating point precision affects results
In the capillary resistance calculation routine, the condition where
Pin and Pout are very similar can occur during low flow situation. To
prevent round errors and large relative fluctuations that do not
influence the model result, set R to 0 when pressure gradient is less
than 1 part per 1,000,000.

Furthermore, avoid the condition of dividing by 0 if the flow
increases on the next iteration(s) and cap.R was set at 0. Instead,
return the pressure gradient as indicator of change.