Fix OpenCL error with Intel driver
While the __read_only and __write_only attributes to data input and
output areas are not really "per spec" and should be ignored if not
used, nVidia driver seem to silently ignore them (maybe uses them?),
AMD driver generates a warning but an Intel driver outright refuses to
compile the code.

Drop the attributes to make Intel compiler happy.