Handle the case when Pal>0 that results in closed capillaries
When capillaries close, they change geometry of the lung that is a
function of Pal and inlet capillary pressure. This may result in
amplifying oscillation in open/closed capillary states as flow remains
constant. To reduce the oscillation, capillaries are forcefully closed
and then a binary search is performed to find a solution where force
closed capillaries are the capillaries that would be closed due to
inlet pressure lower than Pal.

This corner case results in greatly increases number of iterations
required.