Improve efficiency of drawing lung model
Since we no longer use the builtin indexer, drawing speed of the
entire model has increased. But there is still needless calls for
drawing elements that are either too small or are off-screen. This can
be mitigated by hiding elements.

Of course, the problem here is really breaks model-view programming
style. Hiding elements in the model to speed up the view. Ugly hack,
but there is no workaround unless QGraphicsScene gets a user
programmable index.