Clear memory in "unused" portion of OpenCL calculation block
When passing items to OpenCL computational unit, they are arranged in
a matrix NxM elements. But when NxM contains more units than actual
vessels, the remaining items can contain undefined numbers resulting
in possibly undefined behaviour. When run on a CPU, this undefined
behaviour can trigger programs to abort. On GPU, the kernel can run
undetermined amount of time.

To fix this, simply set all data to 0. results in kernel exiting
quickly from this computation and undefined behaviour is avoided.