Replaced Vfrc, Vrv with Vc and Vd, as requested.
"""
I found an article that describes the behavior of the lung using a
continuous exponential equation.  Vm and Vtlc remain the same, instead
of Vrv we enter a value called "c" and instead of Vfrc we enter
another value called "d".  From Vm, Vtlc, c and d, 2 new values are
calculated "a" and "b".  And lung volume equation becomes

 V=Vm+a+b/(1+exp(-(Ptp-c)/d)
 b=Vtlc(1+Exp(c/d))
 a=b/(1+Exp(c/d))

The curve is S shaped, c is the Ptp at which the curve changes between
concave to convex (i.e., the differential of the V equation is 0) and
d is the width of the linear (middle portion of the curve).
"""