Starling resistor happens when Ptm<0, so calculate it properly
Starling resistor (fluttering vessels), only happens when Ptm<0. Since
we can know Pout and vessel parameters that are constants (Ppl,
perivascular constants, GP), we can calculate the P0 in units of
Pin/Pout at which Starling Resistor occurs. It is assumed that
"insignificant" length of the vessels sustains this fluttering and
laminar flow remains. The resistance of the starling resistor is then
just,

  (Pout - P0) / flow, IFF Pout<P0

Now, perivascular parameters modify the pressure in the form,

  Ptm = P - a - b*exp(c*P + d)

so to solve for P, we need to use the Lambert W functions.

Currently, this is only implemented in CPU computation. GPU remains
ignorant, for now.